Terrain

+++ mono/world/generators/WorldGenerator.cs
@@ -29,7 +29,7 @@ public partial class WorldGenerator : GodotObject
 
     public int GetValueTier(float value)
     {
-        for (var i = 0; i < _nTiers; i++){if (value < (i + 1.0f)/(float) _nTiers){return i;}}
+        for (var i = 0; i < _nTiers; i++){if (value < (i + 1.0f)/ _nTiers){return i;}}
 		return _nTiers - 1;
     }
